fix code

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -7,7 +7,7 @@
 */
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t n1 = 0;
+	int n1 = 0;
 
 	if (h == NULL)
 		return (0);
